# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Another one weather api.",
        "title": "Weather API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/cortezzIP/Weather-API/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/weather/current": {
            "get": {
                "description": "Get current weather by location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Get current weather",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Weather"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Current": {
            "type": "object",
            "properties": {
                "temp_c": {
                    "type": "number"
                }
            }
        },
        "model.Location": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Weather": {
            "type": "object",
            "properties": {
                "current": {
                    "$ref": "#/definitions/model.Current"
                },
                "location": {
                    "$ref": "#/definitions/model.Location"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}